refactor : update file structure
