+++ src/components/Image/ImagePreview/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useCallback } from 'react';
 import styled from '@emotion/styled';
 import { ImageCarousel } from 'components/image';
-import { useModal } from 'hooks';
+import { useModal } from 'hooks/common';
 import { Icon } from 'atoms';
 import { faPlus, faTrashAlt } from '@fortawesome/free-solid-svg-icons';
 import { colorCode } from 'common/style/color';
