+++ src/components/MyPage/UpdatePassword/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Input, Button } from 'atoms';
 import { inputId } from 'common/constant/input';
-import { useUpdatePassword } from 'hooks';
+import { useUpdatePassword } from 'hooks/controller';
 import { SAVE_CAPTION } from 'common/constant/string';
 import * as S from '../style';
