+++ src/components/auth/SignUp/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useSignUp } from 'hooks';
+import { useSignUp } from 'hooks/controller';
 import { Input, Button } from 'atoms';
 import { inputId } from 'common/constant/input';
 import {
