+++ src/components/common/Header/index.tsx
@@ -2,7 +2,8 @@ import React from 'react';
 import Logo from 'components/ui/Logo';
 import SearchBar from 'components/common/SearchBar';
 import Navigation from 'components/common/Navigation';
-import { useToggle, useSignOut } from 'hooks';
+import { useSignOut } from 'hooks/controller';
+import { useToggle } from 'hooks/common';
 import { Icon, Anchor, Span } from 'atoms';
 import { colorCode } from 'common/style/color';
 import { useRouter } from 'next/router';
