+++ src/components/common/SearchBar/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useInput from 'hooks/useInput';
+import { useInput } from 'hooks/common';
 import styled from '@emotion/styled';
 import { Icon, Button } from 'atoms';
 import { colorCode } from 'common/style/color';
